+++ x/node/types/errors.go
@@ -22,4 +22,5 @@ var (
 	ErrInsufficientCoin = sdkerrors.Register(ModuleName, 3103, "insufficient coin")
 	ErrInvalidCommitId  = sdkerrors.Register(ModuleName, 3104, "invalid commit")
 	ErrInvalidLastOrder = sdkerrors.Register(ModuleName, 3105, "invalid last order id")
+	ErrInvalidDelegate  = sdkerrors.Register(ModuleName, 3106, "invalid delegate")
 )